flat_crawler/flat_crawler/requirements.txt
psycopg2-binary==2.9.9
Scrapy==2.11.0
